requirements.txt
# =========================================
# RL MARKET MAKER - REQUIREMENTS
# =========================================

# Core RL Framework
gymnasium==0.29.1                # Modern gym replacement
stable-baselines3==2.2.1         # PPO/SAC implementations
sb3-contrib==2.2.1               # Additional algorithms

# Deep Learning
torch==2.1.2                     # PyTorch for neural networks
tensorboard==2.15.1              # Training visualization

# Data Processing
numpy==1.24.3                    # Numerical computing
pandas==2.1.4                    # Data manipulation
scipy==1.11.4                    # Scientific computing

# Market Data & APIs
ccxt==4.2.25                     # Crypto exchange APIs (Binance)
requests==2.31.0                 # HTTP requests
websocket-client==1.7.0          # WebSocket for live data
python-binance==1.0.19           # Binance-specific SDK

# C++ Integration
pybind11==2.11.1                 # Python-C++ bindings
setuptools==69.0.3               # For C++ compilation

# Visualization
matplotlib==3.8.2                # Plotting
seaborn==0.13.1                  # Statistical plots
plotly==5.18.0                   # Interactive plots

# Configuration
pyyaml==6.0.1                    # YAML config files
python-dotenv==1.0.0             # Environment variables

# Logging & Monitoring
loguru==0.7.2                    # Better logging
tqdm==4.66.1                     # Progress bars

# Testing (Optional but recommended)
pytest==7.4.4                    # Unit testing
pytest-cov==4.1.0                # Coverage reports

# Jupyter (For notebooks)
jupyter==1.0.0
ipykernel==6.28.0
ipywidgets==8.1.1

# Performance
numba==0.58.1                    # JIT compilation for Python
joblib==1.3.2                    # Parallel processing

# Data Storage
h5py==3.10.0                     # HDF5 for large datasets
pyarrow==14.0.2                  # Fast columnar data

# Utilities
colorama==0.4.6                  # Colored terminal output
rich==13.7.0                     # Rich text formatting